types: Split types across submodules
